Moved test from SpecmaticJUnitSupportKtTest.kt to SpecmaticJunitSupportTest in the Junit%Support project.
