+++ junit5-support/src/test/resources/spec_with_parameterized_paths.yaml
